feat: add intelligent program (#108)
* feat: add intelligent program

I have added reference to the dishwasher program named "Intelligent" in the UI and with program key Dishcare.Dishwasher.Program.LearningDishwasher in the API.

See also: 
https://github.com/home-assistant/core/issues/155236
https://github.com/home-assistant/core/pull/155243#pullrequestreview-3382346294
https://github.com/MartinHjelmare/aiohomeconnect/issues/71

Note that this is my first commit ever so please bear with me if I made any errors. Thanks for your review.

* chore(pre-commit.ci): auto fixes

* Update name

* moved program up so that the names are still sorted

as requested

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Martin Hjelmare <marhje52@gmail.com>